+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.1] - 2020-10-18
 - Added changelog
 - Added service workers
+- Added routing and navigation
 
 ## [0.0.0] - 2020-10-18
 - init